refactor: marketCarCategory -> marketCarKind로 변경, 스토리북 변경
