--- src/components/market/marketCarCategory/index.ts
@@ -1,1 +1,0 @@
-export { default } from './marketCarCategory';
