+++ src/components/market/marketCarKind/marketCarKind.styled.ts
